+++ frontend/src/hooks/useViewToggle/types.ts
@@ -1,0 +1,6 @@
+export interface ViewToggleOperations {
+  isVisible: boolean;
+  toggle: () => void;
+  show: () => void;
+  hide: () => void;
+}
